bug fixed. Bug: Create entry -> click modify -> approved by approver -> can save by maker -> status become pending

+++ ExpenseProcessingSystem/ExpenseProcessingSystem/Controllers/HomeController.cs
@@ -33,7 +33,7 @@ namespace ExpenseProcessingSystem.Controllers
     [ScreenFltr]
     public class HomeController : Controller
     {
-        private readonly int pageSize = 2;
+        private readonly int pageSize = 30;
         private readonly IHttpContextAccessor _httpContextAccessor;
         private readonly EPSDbContext _context;
         private readonly GOExpressContext _GOContext;
@@ -1290,26 +1290,47 @@ namespace ExpenseProcessingSystem.Controllers
             }
             else
             {
-                if (_service.deleteExpenseEntry(EntryCVViewModelList.entryID))
+                List<int> EditableStatus = new List<int>{
+                    GlobalSystemValues.STATUS_PENDING,
+                    GlobalSystemValues.STATUS_REJECTED
+                };
+                var currentStat = _service.GetCurrentEntryStatus(EntryCVViewModelList.entryID);
+                if (EditableStatus.Contains(currentStat))
                 {
-                    id = _service.addExpense_CV(EntryCVViewModelList, int.Parse(GetUserID()), GlobalSystemValues.TYPE_CV);
+                    if (_service.deleteExpenseEntry(EntryCVViewModelList.entryID))
+                    {
+                        id = _service.addExpense_CV(EntryCVViewModelList, int.Parse(GetUserID()), GlobalSystemValues.TYPE_CV);
+                        var makerId = _context.ExpenseEntry.FirstOrDefault(x => x.Expense_ID == id).Expense_Creator_ID;
+                        //----------------------------- NOTIF----------------------------------
+                        _service.insertIntoNotif(int.Parse(userId), GlobalSystemValues.TYPE_CV, GlobalSystemValues.STATUS_EDIT, makerId);
+                        //----------------------------- NOTIF----------------------------------
+                    }
+                }
+                else
+                {
+                    id = EntryCVViewModelList.entryID;
+                    GlobalSystemValues.MESSAGE = GlobalSystemValues.MESSAGE1;
                 }
             }
 
             ModelState.Clear();
-            if (id > -1) {
-                cvList = _service.getExpense(id);
-                List<SelectList> listOfSysVals = _service.getEntrySystemVals();
-                cvList.systemValues.vendors = listOfSysVals[GlobalSystemValues.SELECT_LIST_VENDOR];
-                cvList.systemValues.dept = listOfSysVals[GlobalSystemValues.SELECT_LIST_DEPARTMENT];
-                cvList.systemValues.currency = listOfSysVals[GlobalSystemValues.SELECT_LIST_CURRENCY];
-                cvList.systemValues.ewt = listOfSysVals[GlobalSystemValues.SELECT_LIST_TAXRATE];
-                cvList.systemValues.acc = _service.getAccDetailsEntry();
-                cvList.systemValues.vat = _service.getVendorVat(cvList.systemValues.vendors.DataValueField[0]);
-                ViewBag.Status = cvList.status;
-            }
+            //if (id > -1) {
+            //    cvList = _service.getExpense(id);
+            //    List<SelectList> listOfSysVals = _service.getEntrySystemVals();
+            //    cvList.systemValues.vendors = listOfSysVals[GlobalSystemValues.SELECT_LIST_VENDOR];
+            //    cvList.systemValues.dept = listOfSysVals[GlobalSystemValues.SELECT_LIST_DEPARTMENT];
+            //    cvList.systemValues.currency = listOfSysVals[GlobalSystemValues.SELECT_LIST_CURRENCY];
+            //    cvList.systemValues.ewt = listOfSysVals[GlobalSystemValues.SELECT_LIST_TAXRATE];
+            //    cvList.systemValues.acc = _service.getAccDetailsEntry();
+            //    cvList.systemValues.vat = _service.getVendorVat(cvList.systemValues.vendors.DataValueField[0]);
+            //    ViewBag.Status = cvList.status;
+            //}
 
-            return View("Entry_CV_ReadOnly", cvList);
+            TempData["entryIDAddtoView"] = id;
+
+            return RedirectToAction("View_CV", "Home");
+
+            //return View("Entry_CV_ReadOnly", cvList);
         }
 
         [OnlineUserCheck]
@@ -1420,6 +1441,16 @@ namespace ExpenseProcessingSystem.Controllers
         {
             var userId = GetUserID();
 
+            if (entryID == 0 && TempData["entryIDAddtoView"] != null)
+            {
+                entryID = (int)TempData["entryIDAddtoView"];
+                TempData.Keep();
+            }
+            else
+            {
+                TempData.Remove("entryIDAddtoView");
+            }
+
             EntryCVViewModelList cvList;
             cvList = _service.getExpense(entryID);
             List<SelectList> listOfSysVals = _service.getEntrySystemVals();
@@ -1453,6 +1484,12 @@ namespace ExpenseProcessingSystem.Controllers
                 }
             }
             cvList.birForms.AddRange(birForms);
+
+            if (!String.IsNullOrEmpty(GlobalSystemValues.MESSAGE))
+            {
+                ViewData["MESSAGE"] = GlobalSystemValues.MESSAGE;
+                GlobalSystemValues.MESSAGE = "";
+            }
 
             return View("Entry_CV_ReadOnly", cvList);
         }
@@ -1513,13 +1550,26 @@ namespace ExpenseProcessingSystem.Controllers
             EntryDDVViewModelList ddvList = new EntryDDVViewModelList();
             int id = _service.addExpense_DDV(EntryDDVViewModelList, int.Parse(GetUserID()), GlobalSystemValues.TYPE_DDV);
             ModelState.Clear();
-            return RedirectToAction("View_DDV", "Home", new { entryID = id });
+            //return RedirectToAction("View_DDV", "Home", new { entryID = id });
+            TempData["entryIDAddtoView"] = id;
+
+            return RedirectToAction("View_DDV", "Home");
         }
         [OnlineUserCheck]
         [NonAdminRoleCheck]
         public IActionResult View_DDV(int entryID)
         {
             var userId = GetUserID();
+
+            if (entryID == 0 && TempData["entryIDAddtoView"] != null)
+            {
+                entryID = (int)TempData["entryIDAddtoView"];
+                TempData.Keep();
+            }
+            else
+            {
+                TempData.Remove("entryIDAddtoView");
+            }
 
             EntryDDVViewModelList ddvList = _service.getExpenseDDV(entryID);
             ddvList = PopulateEntry((EntryDDVViewModelList)ddvList);
@@ -1548,6 +1598,13 @@ namespace ExpenseProcessingSystem.Controllers
                 }
             }
             ddvList.birForms.AddRange(birForms);
+
+            if (!String.IsNullOrEmpty(GlobalSystemValues.MESSAGE))
+            {
+                ViewData["MESSAGE"] = GlobalSystemValues.MESSAGE;
+                GlobalSystemValues.MESSAGE = "";
+            }
+
             return View("Entry_DDV_ReadOnly", ddvList);
         }
         [OnlineUserCheck]
@@ -1708,13 +1765,26 @@ namespace ExpenseProcessingSystem.Controllers
             }
             else
             {
-                if (_service.deleteExpenseEntry(EntryCVViewModelList.entryID))
+                List<int> EditableStatus = new List<int>{
+                    GlobalSystemValues.STATUS_PENDING,
+                    GlobalSystemValues.STATUS_REJECTED
+                };
+                var currentStat = _service.GetCurrentEntryStatus(EntryCVViewModelList.entryID);
+                if (EditableStatus.Contains(currentStat))
                 {
-                    id = _service.addExpense_CV(EntryCVViewModelList, int.Parse(GetUserID()), GlobalSystemValues.TYPE_PC);
-                    var makerId = _context.ExpenseEntry.FirstOrDefault(x => x.Expense_ID == id).Expense_Creator_ID;
-                    //----------------------------- NOTIF----------------------------------
-                    _service.insertIntoNotif(int.Parse(userId), GlobalSystemValues.TYPE_PC, GlobalSystemValues.STATUS_EDIT, makerId);
-                    //----------------------------- NOTIF----------------------------------
+                    if (_service.deleteExpenseEntry(EntryCVViewModelList.entryID))
+                    {
+                        id = _service.addExpense_CV(EntryCVViewModelList, int.Parse(GetUserID()), GlobalSystemValues.TYPE_PC);
+                        var makerId = _context.ExpenseEntry.FirstOrDefault(x => x.Expense_ID == id).Expense_Creator_ID;
+                        //----------------------------- NOTIF----------------------------------
+                        _service.insertIntoNotif(int.Parse(userId), GlobalSystemValues.TYPE_PC, GlobalSystemValues.STATUS_EDIT, makerId);
+                        //----------------------------- NOTIF----------------------------------
+                    }
+                }
+                else
+                {
+                    id = EntryCVViewModelList.entryID;
+                    GlobalSystemValues.MESSAGE = GlobalSystemValues.MESSAGE1;
                 }
             }
 
@@ -1962,6 +2032,11 @@ namespace ExpenseProcessingSystem.Controllers
             }
             pcvList.birForms.AddRange(birForms);
 
+            if (!String.IsNullOrEmpty(GlobalSystemValues.MESSAGE))
+            {
+                ViewData["MESSAGE"] = GlobalSystemValues.MESSAGE;
+                GlobalSystemValues.MESSAGE = "";
+            }
 
             return View("Entry_PCV_ReadOnly", pcvList);
         }
@@ -2023,13 +2098,26 @@ namespace ExpenseProcessingSystem.Controllers
             }
             else
             {
-                if (_service.deleteExpenseEntry(EntryCVViewModelList.entryID))
+                List<int> EditableStatus = new List<int>{
+                    GlobalSystemValues.STATUS_PENDING,
+                    GlobalSystemValues.STATUS_REJECTED
+                };
+                var currentStat = _service.GetCurrentEntryStatus(EntryCVViewModelList.entryID);
+                if (EditableStatus.Contains(currentStat))
                 {
-                    id = _service.addExpense_CV(EntryCVViewModelList, int.Parse(GetUserID()), GlobalSystemValues.TYPE_SS);
-                    var makerId = _context.ExpenseEntry.FirstOrDefault(x => x.Expense_ID == id).Expense_Creator_ID;
-                    //----------------------------- NOTIF----------------------------------
-                    _service.insertIntoNotif(int.Parse(userId), GlobalSystemValues.TYPE_SS, GlobalSystemValues.STATUS_EDIT, makerId);
-                    //----------------------------- NOTIF----------------------------------
+                    if (_service.deleteExpenseEntry(EntryCVViewModelList.entryID))
+                    {
+                        id = _service.addExpense_CV(EntryCVViewModelList, int.Parse(GetUserID()), GlobalSystemValues.TYPE_SS);
+                        var makerId = _context.ExpenseEntry.FirstOrDefault(x => x.Expense_ID == id).Expense_Creator_ID;
+                        //----------------------------- NOTIF----------------------------------
+                        _service.insertIntoNotif(int.Parse(userId), GlobalSystemValues.TYPE_SS, GlobalSystemValues.STATUS_EDIT, makerId);
+                        //----------------------------- NOTIF----------------------------------
+                    }
+                }
+                else
+                {
+                    id = EntryCVViewModelList.entryID;
+                    GlobalSystemValues.MESSAGE = GlobalSystemValues.MESSAGE1;
                 }
             }
 
@@ -2305,6 +2393,12 @@ namespace ExpenseProcessingSystem.Controllers
             }
             ssList.birForms.AddRange(birForms);
 
+            if (!String.IsNullOrEmpty(GlobalSystemValues.MESSAGE))
+            {
+                ViewData["MESSAGE"] = GlobalSystemValues.MESSAGE;
+                GlobalSystemValues.MESSAGE = "";
+            }
+
             return View("Entry_SS_ReadOnly", ssList);
         }
 
@@ -2376,13 +2470,26 @@ namespace ExpenseProcessingSystem.Controllers
             //EntryNCViewModelList ncList = new EntryNCViewModelList();
             int id = _service.addExpense_NC(EntryNCViewModelList, int.Parse(GetUserID()), GlobalSystemValues.TYPE_NC);
             ModelState.Clear();
-            return RedirectToAction("View_NC", "Home", new { entryID = id });
+            //return RedirectToAction("View_NC", "Home", new { entryID = id });
+            TempData["entryIDAddtoView"] = id;
+
+            return RedirectToAction("View_NC", "Home");
         }
         [OnlineUserCheck]
         [NonAdminRoleCheck]
         public IActionResult View_NC(int entryID)
         {
             var userId = GetUserID();
+
+            if (entryID == 0 && TempData["entryIDAddtoView"] != null)
+            {
+                entryID = (int)TempData["entryIDAddtoView"];
+                TempData.Keep();
+            }
+            else
+            {
+                TempData.Remove("entryIDAddtoView");
+            }
 
             EntryNCViewModelList ncList = _service.getExpenseNC(entryID);
             ncList = PopulateEntryNC(ncList);
@@ -2448,6 +2555,13 @@ namespace ExpenseProcessingSystem.Controllers
                 }
             }
             ncList.birForms.AddRange(birForms);
+
+            if (!String.IsNullOrEmpty(GlobalSystemValues.MESSAGE))
+            {
+                ViewData["MESSAGE"] = GlobalSystemValues.MESSAGE;
+                GlobalSystemValues.MESSAGE = "";
+            }
+
             return View("Entry_NC_ReadOnly", ncList);
         }
         [OnlineUserCheck]
@@ -2839,13 +2953,26 @@ namespace ExpenseProcessingSystem.Controllers
             }
             else
             {
-                if (_service.deleteLiquidationEntry(vm.entryID))
+                List<int> EditableStatus = new List<int>{
+                    GlobalSystemValues.STATUS_PENDING,
+                    GlobalSystemValues.STATUS_REJECTED
+                };
+                var currentStat = _service.getCurrentLiquidationStatus(vm.entryID);
+                if (EditableStatus.Contains(currentStat))
                 {
-                    id = _service.addLiquidationDetail(vm, int.Parse(GetUserID()), exist);
-                    var makerId = _context.LiquidationEntryDetails.FirstOrDefault(x => x.ExpenseEntryModel.Expense_ID == id).Liq_Created_UserID;
-                    //----------------------------- NOTIF----------------------------------
-                    _service.insertIntoNotif(int.Parse(userId), GlobalSystemValues.TYPE_SS, GlobalSystemValues.STATUS_EDIT, makerId);
-                    //----------------------------- NOTIF----------------------------------
+                    if (_service.deleteLiquidationEntry(vm.entryID))
+                    {
+                        id = _service.addLiquidationDetail(vm, int.Parse(GetUserID()), exist);
+                        var makerId = _context.LiquidationEntryDetails.FirstOrDefault(x => x.ExpenseEntryModel.Expense_ID == id).Liq_Created_UserID;
+                        //----------------------------- NOTIF----------------------------------
+                        _service.insertIntoNotif(int.Parse(userId), GlobalSystemValues.TYPE_SS, GlobalSystemValues.STATUS_EDIT, makerId);
+                        //----------------------------- NOTIF----------------------------------
+                    }
+                }
+                else
+                {
+                    id = vm.entryID;
+                    GlobalSystemValues.MESSAGE = GlobalSystemValues.MESSAGE1;
                 }
             }
 
@@ -2924,6 +3051,12 @@ namespace ExpenseProcessingSystem.Controllers
             if(birForms.Count() > 0)
             {
                 ssList.birForms.AddRange(birForms);
+            }
+
+            if (!String.IsNullOrEmpty(GlobalSystemValues.MESSAGE))
+            {
+                ViewData["MESSAGE"] = GlobalSystemValues.MESSAGE;
+                GlobalSystemValues.MESSAGE = "";
             }
 
             return View("Liquidation_SS_ReadOnly", ssList);
